refactor(domain): userディレクトリをaccountへリネームしアカウント集約を表す
認証ユーザーを連想させる`user`から、集約の実体(ActiveUser/AccountUseCase)に
沿う`account`へ変更し、exports(./account)とimportパスを一括更新した。

Closes #951
