+++ application/packages/domain/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "exports": {
     "./article": "./src/article/index.ts",
-    "./user": "./src/user/index.ts",
+    "./account": "./src/account/index.ts",
     "./*": "./src/*.ts"
   },
   "scripts": {
