+++ application/packages/domain/src/account/schema/auth-schema.ts
